fix layout of the exchange table, move data for the exchange table to MoneyExchangeTable.jsx file, add SideMenu component to App.js
